requirements.txt
fastapi==0.95.0
geopy==2.3.0
gunicorn==20.1.0
joblib==1.2.0
mlflow==2.2.2
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1
xgboost==1.7.4
psycopg2==2.9.5
python-dotenv==1.0.0
uvicorn==0.21.1